Midstream refactoring on clickhouse index file
